requirements.txt
# Core dependencies
google-generativeai>=0.3.0
openai>=1.0.0

# For async support
aiohttp>=3.9.0

# Optional dependencies for enhanced functionality
python-dotenv>=1.0.0  # For loading environment variables from .env file
